Move exchange-connected tests to failsafe.
