+++ .circleci/config.yml
@@ -34,7 +34,7 @@ jobs:
             echo "Pull request: $CIRCLE_PULL_REQUEST"
             echo " - Username: $CIRCLE_PR_USERNAME"
             echo " - Repository: $CIRCLE_PR_REPONAME"
-            mvn -B -U -Pintegration-test clean install -DskipSlowTests=true -Dskip.surefire.tests=true -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -T 1C
+            mvn -B -U -Pintegration-test clean install -Dskip.failsafe.tests=true -Dskip.surefire.tests=true -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -T 1C
 
       - save_cache:
           paths:
