+++ build.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-mvn clean package -Pproduction -Dmaven.test.skip=true -T 1C
+mvn clean package -Pproduction -Dmaven.test.skip=true -Dskip.failsafe.tests=true -T 1C
