--- orko-db-testsupport/src/main/java/com/gruelbox/orko/db/TestingUtils.java
@@ -1,33 +1,0 @@
-/**
- * Orko
- * Copyright © 2018-2019 Graham Crockford
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.gruelbox.orko.db;
-
-
-import static org.junit.Assume.assumeFalse;
-
-public final class TestingUtils {
-
-  private static final String SKIP_SLOW_TESTS = "skipSlowTests";
-
-  private TestingUtils() {}
-
-  public static void skipIfSlowTestsDisabled() {
-    assumeFalse("true".equals(System.getProperty(SKIP_SLOW_TESTS)));
-  }
-}
